requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
opencv-python>=4.5.0
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Training utilities
tqdm>=4.62.0
tensorboard>=2.8.0

# Configuration
pyyaml>=6.0

# Data handling
pandas>=1.3.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Development tools (optional)
black>=22.0.0
flake8>=4.0.0
mypy>=0.931
isort>=5.10.0
pre-commit>=2.17.0

# Documentation (optional)
sphinx>=4.4.0
sphinx-rtd-theme>=1.0.0
sphinx-autodoc-typehints>=1.17.0